Passed the rest of the flowsContext functions to zustand, removed flowsContext
