+++ src/frontend/src/types/zustand/flowsManager/index.ts
@@ -1,4 +1,4 @@
-import { Node, Edge, Viewport } from "reactflow";
+import { Node, Edge, Viewport, XYPosition } from "reactflow";
 import { FlowType } from "../../flow";
 import { FlowState, FlowsState } from "../../tabs";
 
@@ -17,4 +17,9 @@ export type FlowsManagerStoreType = {
   saveFlow: (flow: FlowType, silent?: boolean) => Promise<void>;
   autoSaveCurrentFlow: (nodes: Node[], edges: Edge[], viewport: Viewport) => void;
   uploadFlows: () => Promise<void>;
+  uploadFlow: ({newProject, file, isComponent, position}: {newProject: boolean, file?: File, isComponent?: boolean, position?: XYPosition}) => Promise<string | never>;
+  addFlow: (newProject: boolean, flow?: FlowType, override?: boolean, position?: XYPosition) => Promise<string | undefined>;
+  deleteComponent: (key: string) => Promise<void>;
+  removeFlow: (id: string) => Promise<void>;
+  saveComponent: (component: any, override: boolean) => Promise<string | undefined>;
 };
